requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv